Refine UI by updating color scheme and margins
The commit involves comprehensive changes in the appearance of the Scenarium Editor. It includes updates to the color scheme, transitioning from `DarkGray` to `DarkerGray` for backdrops. Moreover, the margins of several elements have been adjusted for a more fluid visual appearance. Finally, modifications are done to improve the crispness of border rendering with the