Refactor ScenariumEditor: added pin activation and updated Event-I/O-Pin model
In the Scenarium Editor, we have introduced the handling of pin activation events, making the interface more interactive. We also made significant changes to the Event/Input/Output model, transforming it into a unified Pin model, which simplifies the structure and makes it easier to maintain. Additionally, minor edits were made to the Canvas to improve the handling of mouse interactions.