Add CoreInterop functionality with test method
This commit introduces the CoreInterop functionality, with a new ScenariumCore class and a test method. The native C# interop method was refactored to match the new naming conventions. The dll file has also been added to the .gitignore to avoid being version controlled. A prebuild script has been added as well, to ensure the dll file is copied correctly during build.