Replace `get_func_lib` with `take_func_lib` in various files
The function `get_func_lib` has been replaced by `take_func_lib` in several places across various files in the `graph/src` directory. This change not only modified the function signature but also the way it's used. This was done because taking the function library completely instead of just getting a reference is more suitable for the current design.

Signed-off-by: xxorza <1442523+xorza@users.noreply.github.com>