Add support for nodes in CoreContext
A new method, get_nodes, has been added to CoreContext, as well as a corresponding Node struct to handle nodes. The addition allows the retrieval of node data and information. Furthermore, some changes were made to the existing code, which includes the removal of input and output initializations in the get_funcs method.