Add UUID interoperation utility and refactor for better separation
The UUID interoperation utility was added along with related test cases. Some of the existing code was refactored for better separation. Files related to buffer streams and intercommunication have moved to the 'utils' directory to reflect their general purpose use while ensuring that the related functionality is well-organized. The CMake build file was also updated to include relevant changes.