Add FlatBuffers support and update associated classes
This commit introduces FlatBuffers support and updates several associated classes/functions for readability and efficiency. New dependencies for FlatBuffers and Catch2 are fetched and set up in CMake. A generated file for FlatBuffers is introduced with additional FlatBuffers code in test file. New functionalities are also added to 'NodeController' class. Correspond